requirements.txt
appconfig==0.1
appdirs==1.4.4
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==3.2.19
django-cors-headers==4.1.0
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
mysqlclient==2.1.1
pipreqs==0.4.13
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.3
yarg==0.1.9
